+++ launch/demo.launch
@@ -62,6 +62,6 @@
     <arg name="moveit_warehouse_database_path" value="$(arg db_path)"/>
   </include>
 
-  <node pkg="riptide_moveit_config" type="execute_trajectory.py" name="execute_trajectory" output="screen" />
+  <node pkg="riptide_moveit_config" type="linear_interpolator.py" name="linear_interpolator" output="screen" />
 
 </launch>
