--- src/info.txt
@@ -1,9 +1,0 @@
-
-response written to jointtrajectorypoint
-float64[] positions
-float64[] velocities
-float64[] accelerations
-float64[] effort
-duration time_from_start
-
-dont need to write to effort
